% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RMG.R
\name{dMYZdt.RMG}
\alias{dMYZdt.RMG}
\title{Derivatives for adult mosquitoes}
\usage{
dMYZdt.RMG(t, y, pars, Lambda, kappa)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{Lambda}{emergence rate of adult mosquitoes}

\item{kappa}{net infectiousness of human population}
}
\value{
a [numeric] vector
}
\description{
Implements [dMYZdt] for the RMG ODE model.
}
