% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIR.R
\name{dXdt.SEIRdXdH}
\alias{dXdt.SEIRdXdH}
\title{Derivatives for human population}
\usage{
\method{dXdt}{SEIRdXdH}(t, y, pars, FoI)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{FoI}{vector giving the per-capita force of infection for each strata}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dXdt} for the SEIR model with demography.
}
