% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RMG.R
\name{F_eggs.RMG}
\alias{F_eggs.RMG}
\title{Number of eggs laid by adult mosquitoes}
\usage{
\method{F_eggs}{RMG}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_eggs} for the RMG model.
}
