% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RMG.R
\name{F_fqZ.RMG}
\alias{F_fqZ.RMG}
\title{Blood feeding rate of the infective mosquito population}
\usage{
\method{F_fqZ}{RMG}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_fqZ} for the RMG model.
}
