% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIR.R
\name{HTC.SEIR}
\alias{HTC.SEIR}
\title{Compute the HTC for the SEIR model}
\usage{
\method{HTC}{SEIR}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the SEIR model with demography.
}
