% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SIR.R
\name{HTC.SIR}
\alias{HTC.SIR}
\title{Compute the HTC for the SIR model}
\usage{
\method{HTC}{SIR}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the SIR model with demography.
}
