% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{lines_EIR}
\alias{lines_EIR}
\title{Add lines for the EIR \emph{vs.} time}
\usage{
lines_EIR(terms, model, clrs = "black", llty = 1)
}
\arguments{
\item{terms}{a \link{list} with the outputs of \code{exDE::parse_deout()}}

\item{model}{a \link{list} \emph{e.g.},  generated by \code{exDE::xde_setup()} that defines a model in \code{exDE}}

\item{clrs}{a \link{list} of colors for plotting}

\item{llty}{an \link{integer} (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the EIR \emph{vs.} time
}
