% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MYZ.R
\name{lines_M}
\alias{lines_M}
\title{Basic plotting: add to a plot for mosquito population density}
\usage{
lines_M(MYZ, model, clrs)
}
\arguments{
\item{MYZ}{a \link{list} the outputs of exDE::parse_deout}

\item{model}{a \link{list} specifying the model}

\item{clrs}{a \link{list} of colors}
}
\description{
Basic plotting: add to a plot for mosquito population density
}
