% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-X.R
\name{lines_X}
\alias{lines_X}
\title{Basic plotting: add outputs to a plot for models of human infection and immunity}
\usage{
lines_X(XH, model, clrs = "black")
}
\arguments{
\item{XH}{a \link{list} from exDE::parse_deout}

\item{model}{a \link{list} specifying the model}

\item{clrs}{a \link{list} of colors}
}
\description{
Basic plotting: add outputs to a plot for models of human infection and immunity
}
