% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SIS.R
\name{lines_X.SIS}
\alias{lines_X.SIS}
\title{Add the orbits for the SIS model to a plot for models of human infection and immunity}
\usage{
\method{lines_X}{SIS}(XH, model, clrs = "black")
}
\arguments{
\item{XH}{a [list] from exDE::parse_deout}

\item{model}{a [list] specifying the model}

\item{clrs}{a [list] of colors}
}
\description{
Add the orbits for the SIS model to a plot for models of human infection and immunity
}
