% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RM.R
\name{lines_YZ.RM}
\alias{lines_YZ.RM}
\title{Add the orbits for the SIS model to a plot for models of human infection and immunity}
\usage{
\method{lines_YZ}{RM}(MYZ, model, Yclrs = "purple", Zclrs = "darkred")
}
\arguments{
\item{MYZ}{a \link{list} the output of exDE::parse_deout}

\item{model}{a \link{list} specifying the model}

\item{Yclrs}{a \link{list} of colors}

\item{Zclrs}{a \link{list} of colors}
}
\description{
Add the orbits for the SIS model to a plot for models of human infection and immunity
}
