% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RMG.R
\name{lines_YZ_fracs.RMG}
\alias{lines_YZ_fracs.RMG}
\title{Add lines for the fraction infected and infective for the "RMG" model}
\usage{
\method{lines_YZ_fracs}{RMG}(MYZ, model, Yclrs = "purple", Zclrs = "darkred")
}
\arguments{
\item{MYZ}{a [list] the output of exDE::parse_deout}

\item{model}{a [list] specifying the model}

\item{Yclrs}{a [list] of colors}

\item{Zclrs}{a [list] of colors}
}
\description{
Add lines for the fraction infected and infective for the "RMG" model
}
