% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RMG.R
\name{make_indices_MYZ.RMG}
\alias{make_indices_MYZ.RMG}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
make_indices_MYZ.RMG(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements [make_indices_MYZ] for the RMG model.
}
