% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIR.R
\name{make_indices_X.SEIR}
\alias{make_indices_X.SEIR}
\title{Add indices for human population to parameter list}
\usage{
\method{make_indices_X}{SEIR}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{make_indices_X} for the SEIR model.
}
