% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RMG.R
\name{make_inits_MYZ_RMG}
\alias{make_inits_MYZ_RMG}
\title{Make inits for RMG adult mosquito model}
\usage{
make_inits_MYZ_RMG(pars, U_b0, U_g0, Y_b0, Y_g0, Z_b0, Z_g0)
}
\arguments{
\item{pars}{a \link{list}}

\item{U_b0}{total mosquito density at each patch}

\item{U_g0}{total gravid uninfected mosquito density at each patch}

\item{Y_b0}{infected mosquito density at each patch}

\item{Y_g0}{total gravid infected mosquito density at each patch}

\item{Z_b0}{infectious mosquito density at each patch}

\item{Z_g0}{total gravid infectious mosquito density at each patch}
}
\value{
none
}
\description{
Make inits for RMG adult mosquito model
}
