% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RMG.R
\name{make_MYZinits_RMG}
\alias{make_MYZinits_RMG}
\title{Make inits for RMG adult mosquito model}
\usage{
make_MYZinits_RMG(
  pars,
  MYZopts = list(),
  U0 = 5,
  Gu0 = 1,
  Y0 = 1,
  Gy0 = 1,
  Z0 = 1,
  Gz0 = 1
)
}
\arguments{
\item{pars}{a [list]}

\item{MYZopts}{a [list] of values that overwrites the defaults}

\item{U0}{total mosquito density at each patch}

\item{Y0}{infected mosquito density at each patch}

\item{Z0}{infectious mosquito density at each patch}

\item{P0}{total parous mosquito density at each patch}
}
\value{
a [list]
}
\description{
Make inits for RMG adult mosquito model
}
