% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RMG.R
\name{make_MYZpar_RMG}
\alias{make_MYZpar_RMG}
\title{Make parameters for RMG ODE adult mosquito model}
\usage{
make_MYZpar_RMG(
  pars,
  MYZopts = list(),
  calK,
  solve_as = "dde",
  g = 1/12,
  sigma = 1/8,
  f = 0.5,
  q = 0.95,
  eip = 11,
  nu = 1,
  eggsPerBatch = 60
)
}
\arguments{
\item{pars}{a [list]}

\item{MYZopts}{a [list] of values that overwrites the defaults}

\item{calK}{mosquito dispersal matrix of dimensions `nPatches` by `nPatches`}

\item{solve_as}{is either `ode` to solve as an ode or `dde` to solve as a dde}

\item{g}{mosquito mortality rate}

\item{sigma}{emigration rate}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{eip}{length of extrinsic incubation period}

\item{nu}{oviposition rate, per mosquito}

\item{eggsPerBatch}{eggs laid per oviposition}
}
\value{
a [list]
}
\description{
Make parameters for RMG ODE adult mosquito model
}
