% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIR.R
\name{make_Xinits_SEIR}
\alias{make_Xinits_SEIR}
\title{Make initial values for the SEIR human model, with defaults}
\usage{
make_Xinits_SEIR(
  nStrata,
  Xopts = list(),
  H0 = NULL,
  S0 = NULL,
  E0 = 0,
  I0 = 1,
  R0 = 0
)
}
\arguments{
\item{nStrata}{the number of strata in the model}

\item{Xopts}{a \link{list} to overwrite defaults}

\item{H0}{the initial human population density}

\item{S0}{the initial values of the parameter S}

\item{E0}{the initial values of the parameter E}

\item{I0}{the initial values of the parameter I}

\item{R0}{the initial values of the parameter R}
}
\value{
a \link{list}
}
\description{
Make initial values for the SEIR human model, with defaults
}
