% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SIR.R
\name{make_Xinits_SIR}
\alias{make_Xinits_SIR}
\title{Make initial values for the SIR human model, with defaults}
\usage{
make_Xinits_SIR(nStrata, Xopts = list(), H0 = NULL, S0 = NULL, I0 = 1, R0 = 0)
}
\arguments{
\item{nStrata}{the number of strata in the model}

\item{Xopts}{a \link{list} to overwrite defaults}

\item{H0}{the initial human population density}

\item{S0}{the initial values of the parameter S}

\item{I0}{the initial values of the parameter I}

\item{R0}{the initial values of the parameter R}
}
\value{
a \link{list}
}
\description{
Make initial values for the SIR human model, with defaults
}
