% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-garki.R
\name{make_Xpar_garki}
\alias{make_Xpar_garki}
\title{Make parameters for Garki human model}
\usage{
make_Xpar_garki(
  pars,
  Xopts = list(),
  b = 0.55,
  r1 = 0.0023,
  r2 = 0.023,
  nu = 1/15,
  alpha1 = 0.002,
  alpha2 = 0.00019,
  q1 = 0.7,
  q2 = 0.5,
  q3 = 0.3,
  mu = 1/65/365
)
}
\arguments{
\item{pars}{an \link{list}}

\item{Xopts}{an \link{list}}

\item{b}{transmission probability (efficiency) from mosquito to human}

\item{r1}{a \link{numeric} recovery rate for non-immunes}

\item{r2}{a \link{numeric}recovery rate for partially immmunes}

\item{nu}{a \link{numeric}incubation period}

\item{alpha1}{a \link{numeric} rate of losing infectivity}

\item{alpha2}{a \link{numeric} rate of acquiring immunity}

\item{q1}{a \link{numeric} detection of y1}

\item{q2}{a \link{numeric} detection of y2}

\item{q3}{a \link{numeric} detection of y3}

\item{mu}{a \link{numeric} the death rate}
}
\value{
a \link{list}
}
\description{
Make parameters for Garki human model
}
