% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIR.R
\name{make_Xpar_SEIR}
\alias{make_Xpar_SEIR}
\title{Make parameters for SEIR human model, with defaults}
\usage{
make_Xpar_SEIR(pars, Xopts = list(), b = 0.55, r = 1/180, c = 0.15, tau = 1/5)
}
\arguments{
\item{pars}{a \link{list}}

\item{Xopts}{a \link{list} that could overwrite defaults}

\item{b}{transmission probability (efficiency) from mosquito to human}

\item{r}{recovery rate}

\item{c}{transmission probability (efficiency) from human to mosquito}

\item{tau}{the latent period is 1/tau}
}
\value{
a \link{list}
}
\description{
Make parameters for SEIR human model, with defaults
}
