% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RMG.R
\name{MBionomics.RMG}
\alias{MBionomics.RMG}
\title{Reset bloodfeeding and mortality rates to baseline}
\usage{
\method{MBionomics}{RMG}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a named \link{list}
}
\description{
Implements \link{MBionomics} for the RMG model
}
