% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobwork-package.R
\docType{package}
\name{mobwork-package}
\alias{mobwork}
\alias{mobwork-package}
\title{mobwork: Enhanced functionality for exDE}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dd-harp.github.io/mobwork/}
  \item Report bugs at \url{https://github.com/dd-harp/mobwork/issues}
}

}
\keyword{internal}
