% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RMG.R
\name{parse_deout_MYZ.RMG}
\alias{parse_deout_MYZ.RMG}
\title{Parse the output of deSolve and return variables for the RMG model}
\usage{
parse_deout_MYZ.RMG(varslist, deout, pars)
}
\arguments{
\item{deout}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}
}
\value{
none
}
\description{
Implements [parse_deout_MYZ] for the RMG model
}
