% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{plot_aEIR}
\alias{plot_aEIR}
\title{Basic plotting: plot the true aEIR}
\usage{
plot_aEIR(model, clrs = "black", stable = FALSE)
}
\arguments{
\item{model}{a [list] specifying the model}

\item{clrs}{a [list] of colors}

\item{stable}{[logical]}
}
\description{
Basic plotting: plot the true aEIR
}
