% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MYZ.R
\name{plot_M}
\alias{plot_M}
\title{Basic plotting: plot mosquito population density}
\usage{
plot_M(model, clrs, stable, add)
}
\arguments{
\item{model}{a \link{list} specifying the model}

\item{clrs}{a \link{list} of colors}

\item{stable}{\link{logical}}

\item{add}{\link{logical}}
}
\description{
Basic plotting: plot mosquito population density
}
