% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RM.R
\name{plot_M.RM}
\alias{plot_M.RM}
\title{Basic plotting: plot mosquito population density for the "RM" model}
\usage{
\method{plot_M}{RM}(model, clrs = "darkblue", stable = FALSE, add = FALSE)
}
\arguments{
\item{model}{a \link{list} specifying the model}

\item{clrs}{a \link{list} of colors}

\item{stable}{\link{logical}}

\item{add}{\link{logical}}
}
\description{
Basic plotting: plot mosquito population density for the "RM" model
}
