% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{plot_PR}
\alias{plot_PR}
\title{Basic plotting: plot the true PR.}
\usage{
plot_PR(model, clrs = "black", llty = 1, stable = FALSE, add = FALSE)
}
\arguments{
\item{model}{a \link{list} \emph{e.g.},  generated by \code{exDE::xde_setup()} that defines a model in \code{exDE}}

\item{clrs}{a \link{list} of colors for plotting}

\item{llty}{an \link{integer} (or integers) that specifies \code{lty} for plotting}

\item{stable}{a \link{logical} set to FALSE for \code{orbits} and TRUE for \code{stable_orbits}}

\item{add}{a \link{logical} to plot axes only if FALSE}
}
\description{
Basic plotting: plot the true PR.
}
