% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-X.R
\name{plot_X}
\alias{plot_X}
\title{Basic plotting: default plot for models of human infection and immunity}
\usage{
plot_X(model, clrs, stable)
}
\arguments{
\item{model}{a [list] specifying the model}

\item{clrs}{a [list] of colors}

\item{stable}{[logical]}
}
\description{
Basic plotting: default plot for models of human infection and immunity
}
