% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RM.R
\name{plot_YZ.RM}
\alias{plot_YZ.RM}
\title{Basic plotting: plot mosquito population density for the "RM" model}
\usage{
\method{plot_YZ}{RM}(
  model,
  Yclrs = "purple",
  Zclrs = "darkred",
  stable = FALSE,
  add = FALSE
)
}
\arguments{
\item{model}{a \link{list} specifying the model}

\item{Yclrs}{a \link{list} of colors}

\item{Zclrs}{a \link{list} of colors}

\item{stable}{\link{logical}}

\item{add}{\link{logical}}
}
\description{
Basic plotting: plot mosquito population density for the "RM" model
}
