% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RMG.R
\name{plot_YZ_fracs.RMG}
\alias{plot_YZ_fracs.RMG}
\title{Basic plotting: plot the fraction infected and infective for the "RMG" model}
\usage{
\method{plot_YZ_fracs}{RMG}(
  model,
  Yclrs = "purple",
  Zclrs = "darkred",
  stable = FALSE,
  add = FALSE
)
}
\arguments{
\item{model}{a [list] specifying the model}

\item{Yclrs}{a [list] of colors}

\item{Zclrs}{a [list] of colors}

\item{stable}{[logical]}

\item{add}{[logical]}
}
\description{
Basic plotting: plot the fraction infected and infective for the "RMG" model
}
