% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RMG.R
\name{setup_MYZ.RMG}
\alias{setup_MYZ.RMG}
\title{Setup the RMG model}
\usage{
setup_MYZ.RMG(pars, MYZname, nPatches = 1, MYZopts = list(), calK = diag(1))
}
\arguments{
\item{pars}{a \link{list}}

\item{MYZname}{a \link{character} string}

\item{nPatches}{a \link{numeric} value (an integer), the number of patches}

\item{MYZopts}{a \link{list}}

\item{calK}{is a \link{matrix}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZ} for the RMG model
}
