% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-garki.R
\name{setup_X.garki}
\alias{setup_X.garki}
\title{Setup Xpar.Garki}
\usage{
\method{setup_X}{garki}(pars, Xname, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{Xname}{a \link{character} string}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_X} for the Garki model
}
