% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIR.R
\name{setup_X.SEIR}
\alias{setup_X.SEIR}
\title{Setup Xpar.SEIR}
\usage{
\method{setup_X}{SEIR}(pars, Xname, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{Xname}{a \link{character} string}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_X} for the SEIR model
}
