% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SIR.R
\name{setup_X.SIR}
\alias{setup_X.SIR}
\title{Setup Xpar.SIR}
\usage{
\method{setup_X}{SIR}(pars, Xname, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{Xname}{a \link{character} string}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_X} for the SIR model
}
