% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-garki.R
\name{setup_Xinits.garki}
\alias{setup_Xinits.garki}
\title{Setup Xinits.garki}
\usage{
\method{setup_Xinits}{garki}(pars, i, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_Xinits} for the garki model
}
