% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIR.R
\name{setup_Xinits.SEIR}
\alias{setup_Xinits.SEIR}
\title{Setup Xinits.SEIR}
\usage{
\method{setup_Xinits}{SEIR}(pars, i, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_Xinits} for the SEIR model
}
