% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{split_stratum_by_biting}
\alias{split_stratum_by_biting}
\title{Split a stratum into two strata, assigning}
\usage{
split_stratum_by_biting(pars, i, p, fac)
}
\arguments{
\item{pars}{a \link{list} defining a model}

\item{i}{the stratum to split}

\item{p}{the fraction in the higher exposure stratum}

\item{fac}{the factor increase}
}
\value{
pars a list
}
\description{
Split a stratum into two strata, assigning
}
