% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{ssMYZ}
\alias{ssMYZ}
\title{Set up the MYZss object for \code{xde_scaling_lambda}}
\usage{
ssMYZ(model)
}
\arguments{
\item{model}{a list that defines an \code{exDE} model (\emph{e.g.},  generated by \code{xde_setup()})}
}
\description{
This function computes several quantities that are require
}
