% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{xde_eir2pr}
\alias{xde_eir2pr}
\title{Convert eir to pr}
\usage{
xde_eir2pr(eir, model, extend = FALSE)
}
\arguments{
\item{eir}{a \link{vector}}

\item{model}{a \link{list}}

\item{extend}{a \link{logical} option to determine whether to extend outside the range}
}
\description{
Use the outputs$eirpr table to interpolate
}
