% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-garki.R
\name{xde_lines_X.garki}
\alias{xde_lines_X.garki}
\title{Add lines for the density of infected individuals for the Garki model}
\usage{
\method{xde_lines_X}{garki}(XH, pars, clrs = viridisLite::turbo(7), llty = 1)
}
\arguments{
\item{XH}{a \link{list} with the outputs of \code{exDE::parse_deout()}}

\item{pars}{a \link{list} that defines an \code{exDE} model (\emph{e.g.},  generated by \code{xde_setup()})}

\item{clrs}{a \link{character} vector of colors}

\item{llty}{an \link{integer} (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the density of infected individuals for the Garki model
}
