% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SIR.R
\name{xde_lines_X_SIR}
\alias{xde_lines_X_SIR}
\title{Add lines for the density of infected individuals for the SIR model}
\usage{
xde_lines_X_SIR(XH, pars, clrs = c("black", "darkred", "darkblue"), llty = 1)
}
\arguments{
\item{XH}{a list with the outputs of parse_deout_X_SIS}

\item{pars}{a list that defines an \code{exDE} model (\emph{e.g.},  generated by \code{xde_setup()})}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the density of infected individuals for the SIR model
}
