% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{xde_pr2eir}
\alias{xde_pr2eir}
\title{Convert pr to eir}
\usage{
xde_pr2eir(pr, model, extend = FALSE)
}
\arguments{
\item{pr}{a [vector]}

\item{model}{a [list]}

\item{extend}{a [logical] option to determine whether to extend outside the range}
}
\description{
Use the outputs$eirpr table to convert a set of pr values into eir values
}
