% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{xde_pr2lambda}
\alias{xde_pr2lambda}
\title{Convert pr to lambda}
\usage{
xde_pr2lambda(pr, model, extend = FALSE)
}
\arguments{
\item{pr}{a \link{vector}}

\item{model}{a \link{list}}

\item{extend}{a \link{logical} option to determine whether to extend outside the range}
}
\description{
Use the outputs$lambdapr table to convert a set of pr values into lambda values
}
