% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{xde_pr2ni}
\alias{xde_pr2ni}
\title{Convert pr to ni}
\usage{
xde_pr2ni(pr, model, extend = FALSE)
}
\arguments{
\item{pr}{a \link{vector}}

\item{model}{a \link{list}}

\item{extend}{a \link{logical} option to determine whether to extend outside the range}
}
\description{
Use the outputs$eirpr table to interpolate
}
