% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{xde_scaling}
\alias{xde_scaling}
\title{Construct an eirpr object for an arbitary model}
\usage{
xde_scaling(model, N = 25, F_eir = NULL)
}
\arguments{
\item{model}{a \link{list}}

\item{N}{an \link{integer}, the size of the mesh on aEIR}

\item{F_eir}{a \link{function}, the size of the mesh on aEIR}
}
\description{
This takes a model and uses the XH component to define
the eirpr relationship using \code{xde_scaling_eir} then calls \code{xde_scaling_lambda}
}
