% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{xde_scaling_Z}
\alias{xde_scaling_Z}
\title{Compute eir-pr scaling relationships}
\usage{
xde_scaling_Z(model, N = 25)
}
\arguments{
\item{model}{a \link{list}}

\item{N}{an \link{integer}, the size of the mesh on aEIR}
}
\description{
This function computes average annual values for the eir, the pr, and other
interesting terms and returns a table. It is computed for a model of class "human"
}
