% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RMG.R
\name{get_inits_MYZ.RMG}
\alias{get_inits_MYZ.RMG}
\title{Return initial values as a vector}
\usage{
get_inits_MYZ.RMG(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements [get_inits_MYZ] for the RMG model.
}
