% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-garki.R
\name{get_inits_X.garki}
\alias{get_inits_X.garki}
\title{Return initial values as a vector for the Garki model}
\usage{
\method{get_inits_X}{garki}(pars)
}
\arguments{
\item{pars}{a named \link{list}}
}
\value{
a named \link{list}
}
\description{
This method dispatches on the type of \code{pars$Xpar}.
}
