% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-garki.R
\name{HTC.garki}
\alias{HTC.garki}
\title{Compute the HTC for the garki model}
\usage{
\method{HTC}{garki}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the garki model
}
