% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{lines_aEIR}
\alias{lines_aEIR}
\title{Add the orbits for the SIS model to a plot for models of human infection and immunity}
\usage{
lines_aEIR(terms, model, clrs = "black")
}
\arguments{
\item{terms}{a [list] with the outputs of exDE::parse_deout}

\item{model}{an [list]}

\item{clrs}{a [list]}
}
\description{
Add the orbits for the SIS model to a plot for models of human infection and immunity
}
