% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{lines_EIR}
\alias{lines_EIR}
\title{Add lines for the true EIR}
\usage{
lines_EIR(terms, model, clrs = "black")
}
\arguments{
\item{terms}{a \link{list} with the outputs of exDE::parse_deout}

\item{model}{an \link{list}}

\item{clrs}{a \link{list}}
}
\description{
Add lines for the true EIR
}
