% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{lines_eirpr}
\alias{lines_eirpr}
\title{Add lines for an eir-pr scaling relationship}
\usage{
lines_eirpr(model, clrs = "black", llty = 1)
}
\arguments{
\item{model}{a \link{list} \emph{e.g.},  generated by \code{exDE::xde_setup()} that defines a model in \code{exDE}}

\item{clrs}{a \link{list} of colors for plotting}

\item{llty}{a \link{list}}
}
\description{
Add lines for an eir-pr scaling relationship
}
