% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{lines_eirpr}
\alias{lines_eirpr}
\title{Add lines for the EIR vs. the PR}
\usage{
lines_eirpr(model, clrs = "black", llty = 1)
}
\arguments{
\item{model}{an \link{list}}

\item{clrs}{a \link{list}}

\item{llty}{a \link{list}}
}
\description{
Add lines for the EIR vs. the PR
}
