% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{lines_eirVpr}
\alias{lines_eirVpr}
\title{Add lines for the EIR vs. the PR}
\usage{
lines_eirVpr(model, stable = FALSE, clrs = "black", llty = 1)
}
\arguments{
\item{model}{an [list]}

\item{stable}{a [logical]}

\item{clrs}{a [list]}

\item{llty}{a [list]}
}
\description{
Add lines for the EIR vs. the PR
}
