% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{lines_eirVpr}
\alias{lines_eirVpr}
\title{Add lines for the EIR(t) \emph{vs.} the PR(t)}
\usage{
lines_eirVpr(terms, pars, clrs = "black", llty = 1)
}
\arguments{
\item{terms}{a \link{list} with the outputs of \code{exDE::parse_deout()}}

\item{pars}{a \link{list} that defines an \code{exDE} model (\emph{e.g.},  generated by \code{exDE::xde_setup()})}

\item{clrs}{a \link{character} vector of colors}

\item{llty}{an \link{integer} (or integers) that specifies \code{lty} for plotting}
}
\description{
Add lines for the EIR(t) \emph{vs.} the PR(t)
}
