% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{lines_eirVpr}
\alias{lines_eirVpr}
\title{Add lines for the EIR(t) \emph{vs.} the PR(t)}
\usage{
lines_eirVpr(eir, pr, nStrata, clrs = "black", llty = 1)
}
\arguments{
\item{eir}{the daily EIR}

\item{pr}{the parasite rate}

\item{nStrata}{the number of population strata}

\item{clrs}{a \link{character} vector of colors}

\item{llty}{an \link{integer} (or integers) that specifies \code{lty} for plotting}
}
\description{
Add lines for the EIR(t) \emph{vs.} the PR(t)
}
