% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RMG.R
\name{lines_M.RMG}
\alias{lines_M.RMG}
\title{Add the orbits for the SIS model to a plot for models of human infection and immunity}
\usage{
\method{lines_M}{RMG}(MYZ, model, clrs = "darkblue")
}
\arguments{
\item{MYZ}{a \link{list} the outputs of exDE::parse_deout}

\item{model}{a \link{list} specifying the model}

\item{clrs}{a \link{list} of colors}
}
\description{
Add the orbits for the SIS model to a plot for models of human infection and immunity
}
