% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{lines_PR}
\alias{lines_PR}
\title{Add the orbits for the SIS model to a plot for models of human infection and immunity}
\usage{
lines_PR(terms, model, clrs = "black", llty = 1)
}
\arguments{
\item{terms}{a \link{list} with the outputs of \code{exDE::parse_deout()}}

\item{model}{a \link{list} \emph{e.g.},  generated by \code{exDE::xde_setup()} that defines a model in \code{exDE}}

\item{clrs}{a \link{list} of colors for plotting}

\item{llty}{an \link{integer} (or integers) that specifies \code{lty} for plotting}
}
\description{
Add the orbits for the SIS model to a plot for models of human infection and immunity
}
