% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MYZ.R
\name{lines_YZ}
\alias{lines_YZ}
\title{Basic plotting: add lines for the density of infected and infective mosquitoes}
\usage{
lines_YZ(MYZ, model, Yclrs, Zclrs)
}
\arguments{
\item{MYZ}{a [list] the output of exDE::parse_deout}

\item{model}{a [list] specifying the model}

\item{Yclrs}{a [list] of colors}

\item{Zclrs}{a [list] of colors}
}
\description{
Basic plotting: add lines for the density of infected and infective mosquitoes
}
