% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MYZ.R
\name{lines_YZ_fracs}
\alias{lines_YZ_fracs}
\title{Add lines for the fraction of mosquitoes infected and infective}
\usage{
lines_YZ_fracs(MYZ, model, Yclrs, Zclrs)
}
\arguments{
\item{MYZ}{a \link{list} the output of exDE::parse_deout}

\item{model}{a \link{list} specifying the model}

\item{Yclrs}{a \link{list} of colors}

\item{Zclrs}{a \link{list} of colors}
}
\description{
Add lines for the fraction of mosquitoes infected and infective
}
