% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-garki.R
\name{make_indices_X.garki}
\alias{make_indices_X.garki}
\title{Add indices for human population to parameter list}
\usage{
\method{make_indices_X}{garki}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements \link{make_indices_X} for the Garki model.
}
