% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RMG.R
\name{make_inits_MYZ_RMG}
\alias{make_inits_MYZ_RMG}
\title{Make inits for RMG adult mosquito model}
\usage{
make_inits_MYZ_RMG(pars, U0, Gu0, Y0, Gy0, Z0, Gz0)
}
\arguments{
\item{pars}{a \link{list}}

\item{U0}{total mosquito density at each patch}

\item{Gu0}{total gravid uninfected mosquito density at each patch}

\item{Y0}{infected mosquito density at each patch}

\item{Gy0}{total gravid infected mosquito density at each patch}

\item{Z0}{infectious mosquito density at each patch}

\item{Gz0}{total gravid infectious mosquito density at each patch}
}
\value{
none
}
\description{
Make inits for RMG adult mosquito model
}
