% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RMG.R
\name{make_MYZinits_RMG}
\alias{make_MYZinits_RMG}
\title{Make inits for RMG adult mosquito model}
\usage{
make_MYZinits_RMG(
  nPatches,
  MYZopts = list(),
  U_b0 = 5,
  U_g0 = 1,
  Y_b0 = 1,
  Y_g0 = 1,
  Z_b0 = 1,
  Z_g0 = 1
)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{U_b0}{total uninfected, not gravid mosquito density at each patch}

\item{U_g0}{total uninfected, gravid mosquito density at each patch}

\item{Y_b0}{infected, not gravid mosquito density at each patch}

\item{Y_g0}{infected, gravid mosquito density at each patch}

\item{Z_b0}{infectious, not gravid mosquito density at each patch}

\item{Z_g0}{infectious, gravid mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for RMG adult mosquito model
}
