% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIR.R
\name{make_Xinits_SEIR}
\alias{make_Xinits_SEIR}
\title{Make initial values for the SEIR human model, with defaults}
\usage{
make_Xinits_SEIR(pars, Xopts = list(), E0 = 1, I0 = 0, R0 = 1)
}
\arguments{
\item{pars}{a \link{list}}

\item{Xopts}{a \link{list} to overwrite defaults}

\item{E0}{the initial values of the parameter E}

\item{I0}{the initial values of the parameter I}

\item{R0}{the initial values of the parameter R}
}
\value{
a \link{list}
}
\description{
Make initial values for the SEIR human model, with defaults
}
