% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{plot_eirpr}
\alias{plot_eirpr}
\title{Plot the EIR vs. the PR}
\usage{
plot_eirpr(model, clrs = "black", llty = 1)
}
\arguments{
\item{model}{an \link{list}}

\item{clrs}{a \link{list}}

\item{llty}{a \link{list}}
}
\description{
Plot the EIR vs. the PR
}
