% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{plot_eirpr}
\alias{plot_eirpr}
\title{Plot the eir-pr scaling relationship}
\usage{
plot_eirpr(model, clrs = "black", llty = 1)
}
\arguments{
\item{model}{a \link{list} \emph{e.g.},  generated by \code{exDE::xde_setup()} that defines a model in \code{exDE}}

\item{clrs}{a \link{list} of colors for plotting}

\item{llty}{a \link{list}}
}
\description{
Plot the eir-pr scaling relationship
}
