% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{plot_eirpr}
\alias{plot_eirpr}
\title{Plot the EIR vs. the PR}
\usage{
plot_eirpr(model, clrs = "black", stable = FALSE)
}
\arguments{
\item{model}{an [list]}

\item{clrs}{a [list]}
}
\description{
Plot the EIR vs. the PR
}
