% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{plot_eirVpr}
\alias{plot_eirVpr}
\title{Plot the EIR vs. the PR}
\usage{
plot_eirVpr(model, stable = FALSE, clrs = "black", llty = 1)
}
\arguments{
\item{model}{an \link{list}}

\item{stable}{a \link{logical}}

\item{clrs}{a \link{list}}

\item{llty}{a \link{list}}
}
\description{
Plot the EIR vs. the PR
}
