% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MYZ.R
\name{plot_M}
\alias{plot_M}
\title{Basic plotting: plot mosquito population density}
\usage{
plot_M(model, clrs, stable)
}
\arguments{
\item{model}{a [list] specifying the model}

\item{clrs}{a [list] of colors}

\item{stable}{[logical]}
}
\description{
Basic plotting: plot mosquito population density
}
