% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RM.R
\name{plot_M.RM}
\alias{plot_M.RM}
\title{Basic plotting: plot mosquito population density for the "RM" model}
\usage{
\method{plot_M}{RM}(model, clrs = "darkblue", stable = FALSE)
}
\arguments{
\item{model}{a [list] specifying the model}

\item{clrs}{a [list] of colors}

\item{stable}{[logical]}
}
\description{
Basic plotting: plot mosquito population density for the "RM" model
}
