% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RMG.R
\name{plot_M.RMG}
\alias{plot_M.RMG}
\title{Basic plotting: plot mosquito population density for the "RMG" model}
\usage{
\method{plot_M}{RMG}(model, clrs = "darkblue", stable = FALSE, add = FALSE)
}
\arguments{
\item{model}{a [list] specifying the model}

\item{clrs}{a [list] of colors}

\item{stable}{[logical]}

\item{add}{[logical]}
}
\description{
Basic plotting: plot mosquito population density for the "RMG" model
}
