% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{plot_PR}
\alias{plot_PR}
\title{Basic plotting: plot the true PR.}
\usage{
plot_PR(model, clrs = "black", stable = FALSE)
}
\arguments{
\item{model}{a \link{list} specifying the model}

\item{clrs}{a \link{list} of colors}

\item{stable}{\link{logical}}
}
\description{
Basic plotting: plot the true PR.
}
