% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-X.R
\name{plot_X}
\alias{plot_X}
\title{Basic plotting: default plot for models of human infection and immunity}
\usage{
plot_X(model, clrs = "black", stable = FALSE, add = FALSE)
}
\arguments{
\item{model}{a \link{list} specifying the model}

\item{clrs}{a \link{list} of colors}

\item{stable}{\link{logical}}

\item{add}{\link{logical}}
}
\description{
Basic plotting: default plot for models of human infection and immunity
}
