% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SIS.R
\name{plot_X.SIS}
\alias{plot_X.SIS}
\title{Basic plotting: plot the density of infected humans for an SIS model #' #' @inheritParams plot_X}
\usage{
\method{plot_X}{SIS}(model, clrs = "black", stable = FALSE, add = FALSE)
}
\description{
Basic plotting: plot the density of infected humans for an SIS model #' #' @inheritParams plot_X
}
