% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MYZ.R
\name{plot_YZ}
\alias{plot_YZ}
\title{Basic plotting: plot the density of infected and infective mosquitoes}
\usage{
plot_YZ(model, Yclrs, Zclrs, stable, add)
}
\arguments{
\item{model}{a \link{list} specifying the model}

\item{Yclrs}{a \link{list} of colors}

\item{Zclrs}{a \link{list} of colors}

\item{stable}{\link{logical}}

\item{add}{\link{logical}}
}
\description{
Basic plotting: plot the density of infected and infective mosquitoes
}
