% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYZmod-RM.R
\name{plot_YZ.RM}
\alias{plot_YZ.RM}
\title{Basic plotting: plot mosquito population density for the "RM" model}
\usage{
\method{plot_YZ}{RM}(model, Yclrs = "purple", Zclrs = "darkred", stable = FALSE)
}
\arguments{
\item{model}{a [list] specifying the model}

\item{Yclrs}{a [list] of colors}

\item{Zclrs}{a [list] of colors}

\item{stable}{[logical]}
}
\description{
Basic plotting: plot mosquito population density for the "RM" model
}
