% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MYZ.R
\name{plot_YZ_fracs}
\alias{plot_YZ_fracs}
\title{Basic plotting: plot the proportion of mosquitoes that are infected and infective.}
\usage{
plot_YZ_fracs(model, Yclrs, Zclrs, stable)
}
\arguments{
\item{model}{a [list] specifying the model}

\item{Yclrs}{a [list] of colors}

\item{Zclrs}{a [list] of colors}

\item{stable}{[logical]}
}
\description{
Basic plotting: plot the proportion of mosquitoes that are infected and infective.
}
