% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{pr2Lambda}
\alias{pr2Lambda}
\title{Using the eirpr matrix and a RM model, convert pr to Lambda}
\usage{
pr2Lambda(pr, model, constrain = TRUE)
}
\arguments{
\item{pr}{a \link{numeric} vector}

\item{model}{a \link{list}}

\item{constrain}{a \link{logical}}
}
\description{
This takes a model and uses the XH component to define
the eirpr relationship using \code{xde_scaling_eir} then calls \code{xde_scaling_lambda}
}
