% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SIS.R
\name{split_stratum_by_biting.SIS}
\alias{split_stratum_by_biting.SIS}
\title{Split a stratum into two strata, assigning a fraction \code{p} a new biting weight that is multiplied by a factor \code{fac} compared with the old one. The biting weight for the remaining \code{1-p} gets a new factor \code{1/fac}}
\usage{
\method{split_stratum_by_biting}{SIS}(pars, i, p, fac)
}
\arguments{
\item{pars}{a \link{list} defining a model}

\item{i}{the stratum to split}

\item{p}{the fraction in the higher exposure stratum}

\item{fac}{the factor increase}
}
\value{
pars a list

i \link{integer} -- the stratum to split

p \link{numeric} -- the fraction that gets multiplied by \code{fac}

fac \link{numeric} -- the factor
}
\description{
Split a stratum into two strata, assigning a fraction \code{p} a new biting weight that is multiplied by a factor \code{fac} compared with the old one. The biting weight for the remaining \code{1-p} gets a new factor \code{1/fac}
}
