% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-garki.R
\name{update_inits_X.garki}
\alias{update_inits_X.garki}
\title{Update Xinits for the Garki model}
\usage{
\method{update_inits_X}{garki}(pars, y0, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of initial values}

\item{i}{the host species index}
}
\value{
none
}
\description{
Update Xinits for the Garki model
}
