% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmod-SEIR.R
\name{xde_lines_X.SEIR}
\alias{xde_lines_X.SEIR}
\title{Add lines for the density of infected individuals for the SEIR model}
\usage{
\method{xde_lines_X}{SEIR}(
  XH,
  pars,
  clrs = c("black", "darkgreen", "darkred", "darkblue"),
  llty = 1
)
}
\arguments{
\item{XH}{a list with the outputs of \code{exDE::parse_deout()}}

\item{pars}{a list that defines an \code{exDE} model (\emph{e.g.},  generated by \code{xde_setup()})}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the density of infected individuals for the SEIR model
}
