% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{xde_scaling}
\alias{xde_scaling}
\title{Construct an eirpr object for an arbitary model}
\usage{
xde_scaling(model, N = 25, F_eir = NULL)
}
\arguments{
\item{model}{a list that defines an \code{exDE} model (\emph{e.g.},  generated by \code{xde_setup()})}

\item{N}{the size of the mesh on aEIR}

\item{F_eir}{a function to compute the EIR as a function of time}
}
\description{
This takes a model and uses the XH component to define
the eirpr relationship using \code{xde_scaling_eir} then calls \code{xde_scaling_lambda}
}
